requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.11.3
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
altair=5.5.0=pyhd8ed1ab_1
appdirs=1.4.4=pyhd8ed1ab_1
attrs=25.3.0=pyh71513ae_0
aws-c-auth=0.8.7=h7743f02_1
aws-c-cal=0.8.7=h7d555fd_1
aws-c-common=0.12.1=hb9d3cd8_0
aws-c-compression=0.3.1=hcbd9e4e_3
aws-c-event-stream=0.5.4=h286e7e7_3
aws-c-http=0.9.5=hbca0721_0
aws-c-io=0.17.0=ha855f32_8
aws-c-mqtt=0.12.2=hffac463_3
aws-c-s3=0.7.13=h4c9fe3b_3
aws-c-sdkutils=0.2.3=hcbd9e4e_3
aws-checksums=0.2.3=hcbd9e4e_3
aws-crt-cpp=0.31.1=h46b750d_1
aws-sdk-cpp=1.11.510=h1fa5cb7_4
azure-core-cpp=1.14.0=h5cfcd09_0
azure-identity-cpp=1.10.0=h113e628_0
azure-storage-blobs-cpp=12.13.0=h3cf044e_1
azure-storage-common-cpp=12.8.0=h736e048_1
azure-storage-files-datalake-cpp=12.12.0=ha633028_1
beautifulsoup4=4.13.3=pyha770c72_0
blinker=1.9.0=pyhff2d567_0
brotli=1.1.0=hb9d3cd8_2
brotli-bin=1.1.0=hb9d3cd8_2
brotli-python=1.1.0=py310hf71b8c6_2
bzip2=1.0.8=h4bc722e_7
c-ares=1.34.5=hb9d3cd8_0
ca-certificates=2025.1.31=hbcca054_0
cachetools=5.5.2=pyhd8ed1ab_0
certifi=2025.1.31=pyhd8ed1ab_0
cffi=1.17.1=py310h8deb56e_0
charset-normalizer=3.4.1=pyhd8ed1ab_0
click=8.1.8=pyh707e725_0
contourpy=1.3.1=py310h3788b33_0
cycler=0.12.1=pyhd8ed1ab_1
fonttools=4.57.0=py310h89163eb_0
freetype=2.13.3=h48d6fc4_0
frozendict=2.4.6=py310ha75aee5_0
gflags=2.2.2=h5888daf_1005
gitdb=4.0.12=pyhd8ed1ab_0
gitpython=3.1.44=pyhff2d567_0
glog=0.7.1=hbabe93e_0
h2=4.2.0=pyhd8ed1ab_0
hpack=4.1.0=pyhd8ed1ab_0
html5lib=1.1=pyhd8ed1ab_2
hyperframe=6.1.0=pyhd8ed1ab_0
idna=3.10=pyhd8ed1ab_1
importlib-metadata=8.6.1=pyha770c72_0
importlib_resources=6.5.2=pyhd8ed1ab_0
jinja2=3.1.6=pyhd8ed1ab_0
jsonschema=4.23.0=pyhd8ed1ab_1
jsonschema-specifications=2024.10.1=pyhd8ed1ab_1
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.7=py310h3788b33_0
krb5=1.21.3=h659f571_0
lcms2=2.17=h717163a_0
ld_impl_linux-64=2.43=h712a8e2_4
lerc=4.0.0=h27087fc_0
libabseil=20250127.1=cxx17_hbbce691_0
libarrow=19.0.1=hb90904d_7_cpu
libarrow-acero=19.0.1=hcb10f89_7_cpu
libarrow-dataset=19.0.1=hcb10f89_7_cpu
libarrow-substrait=19.0.1=h1bed206_7_cpu
libblas=3.9.0=31_h59b9bed_openblas
libbrotlicommon=1.1.0=hb9d3cd8_2
libbrotlidec=1.1.0=hb9d3cd8_2
libbrotlienc=1.1.0=hb9d3cd8_2
libcblas=3.9.0=31_he106b2a_openblas
libcrc32c=1.1.2=h9c3ff4c_0
libcurl=8.13.0=h332b0f4_0
libdeflate=1.23=h4ddbbb0_0
libedit=3.1.20250104=pl5321h7949ede_0
libev=4.33=hd590300_2
libevent=2.1.12=hf998b51_1
libexpat=2.7.0=h5888daf_0
libffi=3.4.6=h2dba641_1
libgcc=14.2.0=h767d61c_2
libgcc-ng=14.2.0=h69a702a_2
libgfortran=14.2.0=h69a702a_2
libgfortran5=14.2.0=hf1ad2bd_2
libgomp=14.2.0=h767d61c_2
libgoogle-cloud=2.36.0=hc4361e1_1
libgoogle-cloud-storage=2.36.0=h0121fbd_1
libgrpc=1.71.0=he753a82_0
libiconv=1.18=h4ce23a2_1
libjpeg-turbo=3.0.0=hd590300_1
liblapack=3.9.0=31_h7ac8fdf_openblas
liblzma=5.8.1=hb9d3cd8_0
libnghttp2=1.64.0=h161d5f1_0
libnsl=2.0.1=hd590300_0
libopenblas=0.3.29=pthreads_h94d23a6_0
libopentelemetry-cpp=1.20.0=hd1b1c89_0
libopentelemetry-cpp-headers=1.20.0=ha770c72_0
libparquet=19.0.1=h081d1f1_7_cpu
libpng=1.6.47=h943b412_0
libprotobuf=5.29.3=h501fc15_0
libre2-11=2024.07.02=hba17884_3
libsqlite=3.49.1=hee588c1_2
libssh2=1.11.1=hf672d98_0
libstdcxx=14.2.0=h8f9b012_2
libstdcxx-ng=14.2.0=h4852527_2
libthrift=0.21.0=h0e7cc3e_0
libtiff=4.7.0=hd9ff511_3
libutf8proc=2.10.0=h4c51ac1_0
libuuid=2.38.1=h0b41bf4_0
libwebp-base=1.5.0=h851e524_0
libxcb=1.17.0=h8a09558_0
libxcrypt=4.4.36=hd590300_1
libxml2=2.13.7=h81593ed_1
libxslt=1.1.39=h76b75d6_0
libzlib=1.3.1=hb9d3cd8_2
lxml=5.3.2=py310h490dddc_0
lz4-c=1.10.0=h5888daf_1
markupsafe=3.0.2=py310h89163eb_1
matplotlib-base=3.10.1=py310h68603db_0
multitasking=0.0.9=pyhd8ed1ab_0
munkres=1.1.4=pyh9f0ad1d_0
narwhals=1.34.1=pyhd8ed1ab_0
ncurses=6.5=h2d0b736_3
nlohmann_json=3.12.0=h3f2d84a_0
numpy=2.2.4=py310hefbff90_0
openjpeg=2.5.3=h5fbd93e_0
openssl=3.5.0=h7b32b05_0
orc=2.1.1=h17f744e_1
packaging=24.2=pyhd8ed1ab_2
pandas=2.2.3=py310h5eaa309_3
patsy=1.0.1=pyhd8ed1ab_1
peewee=3.17.9=py310h2b6b41d_0
pillow=11.1.0=py310h7e6dc6c_0
pip=25.0.1=pyh8b19718_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_2
platformdirs=4.3.7=pyh29332c3_0
prometheus-cpp=1.3.0=ha5d0236_0
protobuf=5.29.3=py310hcba5963_0
pthread-stubs=0.4=hb9d3cd8_1002
pyarrow=19.0.1=py310hff52083_0
pyarrow-core=19.0.1=py310hac404ae_0_cpu
pycparser=2.22=pyh29332c3_1
pydeck=0.9.1=pyhd8ed1ab_0
pyparsing=3.2.3=pyhd8ed1ab_1
pysocks=1.7.1=pyha55dd90_7
python=3.10.17=hd6af730_0_cpython
python-dateutil=2.9.0.post0=pyhff2d567_1
python-tzdata=2025.2=pyhd8ed1ab_0
python_abi=3.10=6_cp310
pytz=2025.2=pyhd8ed1ab_0
pyyaml=6.0.2=py310h89163eb_2
qhull=2020.2=h434a139_5
re2=2024.07.02=h9925aae_3
readline=8.2=h8c095d6_2
referencing=0.36.2=pyh29332c3_0
requests=2.32.3=pyhd8ed1ab_1
rpds-py=0.24.0=py310hc1293b2_0
s2n=1.5.15=hd830067_0
scipy=1.15.2=py310h1d65ade_0
seaborn=0.13.2=hd8ed1ab_3
seaborn-base=0.13.2=pyhd8ed1ab_3
setuptools=78.1.0=pyhff2d567_0
six=1.17.0=pyhd8ed1ab_0
smmap=5.0.2=pyhd8ed1ab_0
snappy=1.2.1=h8bd8927_1
soupsieve=2.5=pyhd8ed1ab_1
statsmodels=0.14.4=py310hf462985_0
streamlit=1.44.1=pyhd8ed1ab_0
tenacity=9.1.2=pyhd8ed1ab_0
tk=8.6.13=noxft_h4845f30_101
toml=0.10.2=pyhd8ed1ab_1
tornado=6.4.2=py310ha75aee5_0
typing-extensions=4.13.2=h0e9735f_0
typing_extensions=4.13.2=pyh29332c3_0
tzdata=2025b=h78e105d_0
unicodedata2=16.0.0=py310ha75aee5_0
urllib3=2.4.0=pyhd8ed1ab_0
watchdog=6.0.0=py310hff52083_0
webencodings=0.5.1=pyhd8ed1ab_3
wheel=0.45.1=pyhd8ed1ab_1
xorg-libxau=1.0.12=hb9d3cd8_0
xorg-libxdmcp=1.1.5=hb9d3cd8_0
yaml=0.2.5=h7f98852_2
yfinance=0.2.55=pyhd8ed1ab_0
zipp=3.21.0=pyhd8ed1ab_1
zlib=1.3.1=hb9d3cd8_2
zstandard=0.23.0=py310ha75aee5_1
zstd=1.5.7=hb8e6e7a_2
